Задание 5/user_service/requirements.txt
fastapi==0.95.2
uvicorn==0.22.0
sqlalchemy==2.0.15
psycopg2-binary==2.9.6
pydantic[email]==1.10.7
python-dotenv==1.0.0
passlib[bcrypt]==1.7.4
python-jose[cryptography]==3.3.0
python-multipart==0.0.6 
email-validator==1.3.1  
redis==4.3.4
hiredis==2.0.0
psutil==5.9.5